Version == 1.3.3

+++ src/selenium_driverless/webdriver.py
@@ -85,7 +85,10 @@ class Chrome(BaseWebDriver):
         if disconnect_connect:
             warnings.warn("disconnect_connect=True might be buggy")
         self._page_enabled = None
+
         self._global_this = None
+        self._document_node_id_ = None
+
         self._loop: asyncio.AbstractEventLoop or None = None
         self._page_load_timeout: int = 30
         self._script_timeout: int = 30
@@ -193,35 +196,41 @@ class Chrome(BaseWebDriver):
 
         if not self._options.debugger_address:
             from selenium_driverless.utils.utils import random_port
-            port = random_port("localhost")
-            self._options._debugger_address = f"localhost:{port}"
+            port = random_port()
+            self._options._debugger_address = f"127.0.0.1:{port}"
             self._options.add_argument(f"--remote-debugging-port={port}")
         options = capabilities["goog:chromeOptions"]
 
-        path = options["binary"]
-        args = options["args"]
-        cmds = [path, *args]
-        if IS_POSIX:
-            args = " ".join(args)
-            cmds = [f'"{path}" {args}']
-        browser = subprocess.Popen(
-            cmds,
-            stdin=subprocess.PIPE,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            close_fds=IS_POSIX,
-            shell=IS_POSIX
-        )
+        # noinspection PyProtectedMember
+        self._is_remote = self._options._is_remote
+
+        if not self._is_remote:
+            path = options["binary"]
+            args = options["args"]
+            cmds = [path, *args]
+            browser = subprocess.Popen(
+                cmds,
+                stdin=subprocess.PIPE,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                close_fds=IS_POSIX,
+                shell=IS_POSIX
+            )
+
+            # noinspection PyUnboundLocalVariable
+            if port == 0:
+                path = self._options.user_data_dir + "/DevToolsActivePort"
+                while not os.path.isfile(path):
+                    await self.implicitly_wait(0.1)
+                port = int(read(path, sel_root=False).split("\n")[0])
+                self._options.debugger_address = f"127.0.0.1:{port}"
+
         host, port = self._options.debugger_address.split(":")
         port = int(port)
-        if port == 0:
-            path = self._options.user_data_dir + "/DevToolsActivePort"
-            while not os.path.isfile(path):
-                await self.implicitly_wait(0.1)
-            port = int(read(path, sel_root=False).split("\n")[0])
-            self._options.debugger_address = f"localhost:{port}"
-        self._base = await CDPSocket(port=port, host=host, loop=self._loop)
-        self.browser_pid = browser.pid
+        self._base = await CDPSocket(port=port, host=host, loop=self._loop, timeout=30)
+        if not self._is_remote:
+            # noinspection PyUnboundLocalVariable
+            self.browser_pid = browser.pid
         targets = await self._base.targets
         for target in targets:
             if target["type"] == "page":
@@ -234,8 +243,10 @@ class Chrome(BaseWebDriver):
         else:
             self._switch_to = await SwitchTo(driver=self)
 
+        # noinspection PyUnusedLocal
         def clear_global_this(data):
             self._global_this = None
+            self._document_node_id_ = None
 
         await self.add_cdp_listener("Page.loadEventFired", clear_global_this)
 
@@ -276,16 +287,18 @@ class Chrome(BaseWebDriver):
                 raise TimeoutError(f"page didn't load within timeout of {self._page_load_timeout}")
         await get
         self._global_this = None
+        self._document_node_id_ = None
 
     @property
     async def title(self) -> str:
+        # noinspection GrazieInspection
         """Returns the title of the current page.
 
-        :Usage:
-            ::
+                :Usage:
+                    ::
 
-                title = driver.title
-        """
+                        title = driver.title
+                """
         target = await self.current_target
         return target["title"]
 
@@ -335,13 +348,15 @@ class Chrome(BaseWebDriver):
         return res
 
     async def execute_raw_script(self, script: str, *args, await_res: bool = False, serialization: str = None,
-                                 max_depth: int = None, timeout: int = 2, obj_id=None):
+                                 max_depth: int = None, timeout: int = 2, obj_id=None, warn: bool = False):
         """
         example:
         script= "function(...arguments){this.click()}"
         "this" will be the element object
         """
         from selenium_driverless.types import RemoteObject, JSEvalException
+        if warn:
+            warnings.warn("execute_script might be detected", stacklevel=4)
         if not obj_id:
             if not self._global_this:
                 self._global_this = await RemoteObject(driver=self, js="globalThis", check_existence=False)
@@ -372,14 +387,14 @@ class Chrome(BaseWebDriver):
 
     async def execute_script(self, script: str, *args, max_depth: int = 2, serialization: str = None,
                              timeout: int = None,
-                             only_value=True, obj_id=None):
+                             only_value=True, obj_id=None, warn: bool = False):
         """
         exaple: script = "return elem.click()"
         """
         script = f"(function(...arguments){{{script}}})"
         res = await self.execute_raw_script(script, *args, max_depth=max_depth,
                                             serialization=serialization, timeout=timeout,
-                                            await_res=False, obj_id=obj_id)
+                                            await_res=False, obj_id=obj_id, warn=warn)
         if only_value:
             if "value" in res.keys():
                 return res["value"]
@@ -388,14 +403,14 @@ class Chrome(BaseWebDriver):
 
     async def execute_async_script(self, script: str, *args, max_depth: int = 2,
                                    serialization: str = None, timeout: int = 2,
-                                   only_value=True, obj_id=None):
+                                   only_value=True, obj_id=None, warn: bool = False):
         script = """(function(...arguments){
                        const promise = new Promise((resolve, reject) => {
                               arguments.push(resolve)
                         });""" + script + ";return promise})"
         res = await self.execute_raw_script(script, *args, max_depth=max_depth,
                                             serialization=serialization, timeout=timeout,
-                                            await_res=True, obj_id=obj_id)
+                                            await_res=True, obj_id=obj_id, warn=warn)
         if only_value:
             if "value" in res.keys():
                 return res["value"]
@@ -445,28 +460,29 @@ class Chrome(BaseWebDriver):
 
                 driver.quit()
         """
-        import os
-        import shutil
-        # noinspection PyBroadException,PyUnusedLocal
-        try:
+        if not self._is_remote:
+            import os
+            import shutil
+            # noinspection PyBroadException,PyUnusedLocal
             try:
-                await self.base.close()
-                # wait for process to be killed
-                while True:
-                    try:
-                        os.kill(self.browser_pid, 15)
-                    except OSError:
-                        break
-                    await self.implicitly_wait(0.1)
+                try:
+                    await self.base.close()
+                    # wait for process to be killed
+                    while True:
+                        try:
+                            os.kill(self.browser_pid, 15)
+                        except OSError:
+                            break
+                        await self.implicitly_wait(0.1)
 
-                shutil.rmtree(self._options.user_data_dir, ignore_errors=True)
+                    shutil.rmtree(self._options.user_data_dir, ignore_errors=True)
+                finally:
+                    await self.stop_client()
+            except Exception as e:
+                # We don't care about the message because something probably has gone wrong
+                pass
             finally:
-                await self.stop_client()
-        except Exception as e:
-            # We don't care about the message because something probably has gone wrong
-            pass
-        finally:
-            pass  # self.service.stop()
+                pass  # self.service.stop()
 
     @property
     async def targets(self) -> dict:
@@ -662,6 +678,7 @@ class Chrome(BaseWebDriver):
         """
         await self.execute_cdp_cmd("Network.clearBrowserCookies")
 
+    # noinspection GrazieInspection
     async def add_cookie(self, cookie_dict) -> None:
         """Adds a cookie to your current session.
 
@@ -743,16 +760,43 @@ class Chrome(BaseWebDriver):
         self._page_load_timeout = timeouts["page_load"]
         self._script_timeout = timeouts["script"]
 
+    @property
+    async def _document_node_id(self):
+        if not self._document_node_id_:
+            res = await self.execute_cdp_cmd("DOM.getDocument", {"pierce": True})
+            self._document_node_id_ = res["root"]["nodeId"]
+        return self._document_node_id_
+
     # noinspection PyUnusedLocal
     async def find_element(self, by: str, value: str, parent=None):
         if not parent:
-            parent = await WebElement(driver=self, js="document", check_existence=False)
+            parent = await WebElement(driver=self, node_id=await self._document_node_id, check_existence=False, loop=self._loop)
         return await parent.find_element(by=by, value=value)
 
     async def find_elements(self, by: str, value: str, parent=None):
         if not parent:
-            parent = await WebElement(driver=self, js="document", check_existence=False)
+            parent = await WebElement(driver=self, node_id=await self._document_node_id, check_existence=False, loop=self._loop)
         return await parent.find_elements(by=by, value=value)
+
+    async def search_elements(self, query: str):
+        """
+        query:str | Plain text or query selector or XPath search query.
+        """
+        elems = []
+        res = await self.execute_cdp_cmd("DOM.performSearch",
+                                         {"includeUserAgentShadowDOM": True, "query": query})
+        search_id = res["searchId"]
+        elem_count = res["resultCount"]
+
+        res = await self.execute_cdp_cmd("DOM.getSearchResults",
+                                         {"searchId": search_id, "fromIndex": 0, "toIndex": elem_count - 1})
+        for node_id in res["nodeIds"]:
+            if self._loop:
+                elem = await SyncWebElement(driver=self, check_existence=False, node_id=node_id, loop=self._loop)
+            else:
+                elem = await WebElement(driver=self, check_existence=False, node_id=node_id, loop=self._loop)
+            elems.append(elem)
+        return elems
 
     @property
     def capabilities(self) -> dict:
@@ -966,13 +1010,14 @@ class Chrome(BaseWebDriver):
 
     @property
     def orientation(self):
+        # noinspection GrazieInspection
         """Gets the current orientation of the device.
 
-        :Usage:
-            ::
+                :Usage:
+                    ::
 
-                orientation = driver.orientation
-        """
+                        orientation = driver.orientation
+                """
         raise NotImplementedError()
 
     @orientation.setter
